SceneComponent derive / reusable Component derive / SceneComponent: Component
