+++ crates/bevy_ecs/macro_logic/Cargo.toml
@@ -1,20 +1,17 @@
 [package]
-name = "bevy_ecs_macros"
+name = "bevy_ecs_macro_logic"
 version = "0.19.0-dev"
-description = "Bevy ECS Macros"
+description = "Shared Bevy ECS Macro internals"
 edition = "2024"
 license = "MIT OR Apache-2.0"
 
-[lib]
-proc-macro = true
-
 [dependencies]
 bevy_macro_utils = { path = "../../bevy_macro_utils", version = "0.19.0-dev" }
-bevy_ecs_macro_logic = { path = "../macro_logic", version = "0.19.0-dev" }
 
 syn = { version = "2.0.108", features = ["full", "extra-traits"] }
 quote = "1.0"
 proc-macro2 = "1.0"
+
 [lints]
 workspace = true
