+++ crates/bevy_macro_utils/src/lib.rs
@@ -17,6 +17,7 @@ mod label;
 mod member;
 mod parser;
 mod path;
+mod path_type;
 mod result_sifter;
 mod shape;
 mod symbol;
@@ -27,6 +28,7 @@ pub use label::*;
 pub use member::*;
 pub use parser::*;
 pub use path::*;
+pub use path_type::*;
 pub use result_sifter::*;
 pub use shape::*;
 pub use symbol::*;
