+++ crates/bevy_scene/Cargo.toml
@@ -14,9 +14,7 @@ bevy_scene_macros = { path = "macros", version = "0.19.0-dev" }
 bevy_app = { path = "../bevy_app", version = "0.19.0-dev" }
 bevy_asset = { path = "../bevy_asset", version = "0.19.0-dev" }
 bevy_derive = { path = "../bevy_derive", version = "0.19.0-dev" }
-bevy_ecs = { path = "../bevy_ecs", version = "0.19.0-dev", features = [
-  "derive_scene",
-] }
+bevy_ecs = { path = "../bevy_ecs", version = "0.19.0-dev" }
 bevy_log = { path = "../bevy_log", version = "0.19.0-dev" }
 bevy_platform = { path = "../bevy_platform", version = "0.19.0-dev" }
 bevy_reflect = { path = "../bevy_reflect", version = "0.19.0-dev" }
