+++ crates/bevy_scene/macros/Cargo.toml
@@ -13,6 +13,8 @@ proc-macro = true
 
 [dependencies]
 bevy_macro_utils = { path = "../../bevy_macro_utils", version = "0.19.0-dev" }
+bevy_ecs_macro_logic = { path = "../../bevy_ecs/macro_logic", version = "0.19.0-dev" }
+
 syn = { version = "2.0", features = ["full", "extra-traits"] }
 proc-macro2 = "1.0"
 quote = "1.0"
